fixed phone call case handling once and for all.
ultimately learned that the lockscreen is being restored on all timeout sleeps. when the call is then ended, it gets restored. my call end method now handles this by checking for KG, and running the exit (2c secure exit method). 

if the call ends and screen is asleep, lockscreen seems to always exist. so we now always startLock in that case. will replace with the 2.1 screen on check so mediator no longer needs to keep track, as there are times when we get inconsistent accuracy with that method.

also added finish command for back key in settings.